Optimize flat color rendering & gradient rendering
Up until now, we mostly used the `fillColor` method
of `UiRenderer`, which is flexible, but can be
rather slow on old hardware when a large number of
pixels is filled.

This commit dramatically improves the performance on
old hardware by creating a dedicated graphics pipeline
for flat color filling and gradient rendering.